backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.8.0
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx>=0.27.0
starlette>=0.37.0

# Database
motor==3.5.0
pymongo==4.8.0
aiosqlite==0.20.0
sqlalchemy==2.0.31
greenlet==3.0.3
asyncpg==0.29.0
cryptography>=42.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# AI/ML Libraries
transformers==4.42.3
torch>=2.2.0
sentence-transformers==3.0.1
accelerate==0.31.0

# Voice Analysis
librosa==0.10.2
soundfile==0.12.1

# Face Analysis / Vision
opencv-python-headless==4.10.0.84
mediapipe==0.10.14

# Utilities
numpy>=1.26.0
requests==2.32.3
faker==25.8.0

# Free AI Core
faster-whisper==1.0.3
pyttsx3==2.90
chromadb==0.5.4
groq==0.9.0
elevenlabs

# LangChain Stack (v0.2+)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.7
langchain-groq>=0.1.0
langsmith>=0.1.0

# LangGraph (background agents - optional)
langgraph>=0.1.0
